requirements.txt
# Framework Web
Flask==3.0.0
Flask-CORS==4.0.0

# Servidor de producción
gunicorn==21.2.0

# Base de datos
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.0
mysql-connector-python==8.2.0
cryptography==41.0.7

# Variables de entorno
python-dotenv==1.0.0

# Validación
Flask-Marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0

# Seguridad
Flask-Bcrypt==1.0.1
PyJWT==2.8.0

# Utilidades
python-dateutil==2.8.2
